requirements.txt
torchvision==0.21.0
torchaudio==2.6.0
datasets==3.6.0
transformers==4.52.0
accelerate==1.11.0
evaluate==0.4.6
tensorboard==2.19.0
